% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgraphy2.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Creation of Scatter plot using Seaborns ScatterPlot Function}
\usage{
scatterPlot(
  python_path,
  data,
  write_directory,
  figure_name,
  x_variable,
  y_variable,
  graident_variable = FALSE,
  size_variable = FALSE,
  custom_ranking = FALSE,
  style_sheet = FALSE
)
}
\arguments{
\item{python_path}{The path to your python installation}

\item{data}{The path to your data directory}

\item{write_directory}{The path you want to write the plot out to}

\item{figure_name}{The name of the png file you want to use, defaults to GraphyFigure}

\item{x_variable}{The x variable name from your data column header names you want to use as the x variable}

\item{y_variable}{The y variable name from your data column header names you want to use as the y variable}

\item{graident_variable}{The varaible to use to add a graident of colour to the graph, optional}

\item{size_variable}{The varaible to control the size of the points in the scatter, optional}

\item{custom_ranking}{A custom list of rankings to order the variables, optional}
}
\value{
None
}
\description{
Creation of Scatter plot using Seaborns ScatterPlot Function
}
